readme: Restructure into something more coherent
This commit attempts to restructure the README into a format that
introduces the library in a way that new users would consume the
library. I attempted to transition to these points as quickly as
possible. I also moved the goals section to the end of the document
to avoid inundating new users with information.